+++ src/sdk/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Support for working with IListItem attachments [jansenbe - Bert Jansen]
 - Page scheduling support: enable a pages library for page publish scheduling and schedule the publishing of a page [jansenbe - Bert Jansen]
 - IsSubSite methods on IWeb [jansenbe - Bert Jansen]
 - Support for converting the site's datetime to UTC and vice versa [jansenbe - Bert Jansen]